🚀 [Feat]: StepIndicator 재구성
- rename: StepIndicator => StepLink
- 오류가 발생한 수직 정렬을 위해 하나의 component(Link) 내로 위치
- 수직, 수평에 대한 step은 현재 필요없으므로 제거(추후 필요하면 다시)

Bug: #2

<Issue Tracker>
Fixes: #2
